Remove old rules and use new ones
