+++ tests/test_rules.py
@@ -1,269 +1,283 @@
 # -*- coding: utf-8 -*-
+import json
+import logging
 import datetime as dt
+from collections import defaultdict
 
-from doorman.models import Rule
 from doorman.rules import (
-    BaseRule,
-    BlacklistRule,
-    RuleMatch,
-    WhitelistRule,
+    BaseCondition,
+    EqualCondition,
+    Network,
+    RuleInput,
 )
 
 
-class TestBaseRule:
-
-    def test_will_filter_node_name(self, FakeBaseRule):
-        rule = FakeBaseRule(config={'node_name': 'yes'})
-
-        # This should not succeed due to a host_identifier that does not match
-        rule.handle_log_entry({}, {'host_identifier': 'no'})
-        assert len(rule.calls) == 0
-
-        # This should succeed since the host_identifier does match.  Note that
-        # we explicitly don't care about the hostIdentifier value in the query
-        # results.
-        now = dt.datetime.utcnow()
-        rule.handle_log_entry({
-            'data': [
-                {
-                    "diffResults": {
-                        "added": [{'op': 'added'}],
-                        "removed": "",
-                    },
-                    "name": "fake",
-                    "hostIdentifier": "hostname.local",
-                    "calendarTime": "%s %s" % (now.ctime(), "UTC"),
-                    "unixTime": now.strftime('%s')
-                }
-            ]
-        }, {'host_identifier': 'yes'})
-        assert len(rule.calls) == 1
+DUMMY_INPUT = RuleInput(result_log={}, node={})
 
 
-class TestEachResultRule:
-    def setup_method(self, _method):
-        self.now = dt.datetime.utcnow()
-        self.fake_data = {
-            'data': [
-                {
-                    "diffResults": {
-                        "added": [{'op': 'added'}],
-                        "removed": [{'op': 'removed'}],
-                    },
-                    "name": "fake",
-                    "hostIdentifier": "hostname.local",
-                    "calendarTime": "%s %s" % (self.now.ctime(), "UTC"),
-                    "unixTime": self.now.strftime('%s')
-                },
-            ],
+class TestNetwork:
+
+    def test_will_cache_condition_instances(self):
+        class TestCondition(BaseCondition):
+            pass
+
+        network = Network()
+        one = network.make_condition(TestCondition)
+        two = network.make_condition(TestCondition)
+
+        assert one is two
+
+    def test_will_parse_basic(self):
+        query = json.loads("""
+        {
+          "condition": "AND",
+          "rules": [
+            {
+              "id": "column",
+              "field": "column",
+              "type": "string",
+              "input": "text",
+              "operator": "column_equal",
+              "value": [
+                "model_id",
+                "5500"
+              ]
+            }
+          ]
         }
+        """)
 
-    def test_will_filter_action_added(self, FakeEachResultRule):
-        node = {'host_identifier': 'hostname.local'}
-        rule = FakeEachResultRule(action=Rule.ADDED, config={})
-        rule.handle_log_entry(self.fake_data, node)
+        network = Network()
+        network.parse_query(query)
 
-        assert rule.calls == [(Rule.ADDED, {'op': 'added'}, node)]
+        # AND condition, single column condition
+        assert len(network.conditions) == 2
 
-    def test_will_filter_action_removed(self, FakeEachResultRule):
-        node = {'host_identifier': 'hostname.local'}
-        rule = FakeEachResultRule(action=Rule.REMOVED, config={})
-        rule.handle_log_entry(self.fake_data, node)
-
-        assert rule.calls == [(Rule.REMOVED, {'op': 'removed'}, node)]
-
-    def test_will_filter_action_both(self, FakeEachResultRule):
-        node = {'host_identifier': 'hostname.local'}
-        rule = FakeEachResultRule(action=Rule.BOTH, config={})
-        rule.handle_log_entry(self.fake_data, node)
-
-        assert rule.calls == [
-            (Rule.ADDED, {'op': 'added'}, node),
-            (Rule.REMOVED, {'op': 'removed'}, node),
-        ]
-
-    def test_will_filter_query_name(self, FakeEachResultRule):
-        self.fake_data['data'].append({
-            "diffResults": {
-                "added": [{'op': 'added 2'}],
-                "removed": [{'op': 'removed 2'}],
-            },
-            "name": "other",
-            "hostIdentifier": "hostname.local",
-            "calendarTime": "%s %s" % (self.now.ctime(), "UTC"),
-            "unixTime": self.now.strftime('%s')
-        })
-
-        node = {'host_identifier': 'hostname.local'}
-
-        # No filtering
-        rule = FakeEachResultRule(action=Rule.BOTH, config={})
-        rule.handle_log_entry(self.fake_data, node)
-
-        assert rule.calls == [
-            (Rule.ADDED, {'op': 'added'}, node),
-            (Rule.REMOVED, {'op': 'removed'}, node),
-            (Rule.ADDED, {'op': 'added 2'}, node),
-            (Rule.REMOVED, {'op': 'removed 2'}, node),
-        ]
-
-        rule = FakeEachResultRule(action=Rule.BOTH, config={'query_name': 'other'})
-        rule.handle_log_entry(self.fake_data, node)
-        assert rule.calls == [
-            (Rule.ADDED, {'op': 'added 2'}, node),
-            (Rule.REMOVED, {'op': 'removed 2'}, node),
-        ]
-
-
-class TestBlacklistRule:
-
-    def test_will_blacklist(self):
-        now = dt.datetime.utcnow()
-
-        data = [
+    def test_will_reuse_identical_conditions(self):
+        # Operators are equal in each condition
+        query = json.loads("""
+        {
+          "condition": "AND",
+          "rules": [
             {
-              "diffResults": {
-                "added": [
-                  {
-                    "name": "malware",
-                    "path": "/usr/local/bin/malware",
-                    "pid": "12345"
-                  },
-                  {
-                    "name": "legit",
-                    "path": "/usr/local/bin/legit",
-                    "pid": "6789"
-                  },
-                  {
-                    "name": "malware",
-                    "path": "/usr/local/bin/malware",
-                    "pid": "10000"
-                  },
-                ],
-                "removed": "",
-              },
-              "name": "processes",
-              "hostIdentifier": "hostname.local",
-              "calendarTime": "%s %s" % (now.ctime(), "UTC"),
-              "unixTime": now.strftime('%s'),
+              "condition": "AND",
+              "rules": [
+                {
+                  "id": "query_name",
+                  "field": "query_name",
+                  "type": "string",
+                  "input": "text",
+                  "operator": "equal",
+                  "value": "asdf"
+                }
+              ]
             },
-        ]
-        node = {'host_identifier': 'hostname.local'}
-
-        rule = BlacklistRule(0, Rule.BOTH, config={
-            'field_name': 'name',
-            'blacklist': ['malware'],
-        })
-
-        expected1 = RuleMatch(
-            rule_id=0,
-            action='added',
-            node=node,
-            match=data[0]['diffResults']['added'][0],
-        )
-        expected2 = RuleMatch(
-            rule_id=0,
-            action='added',
-            node=node,
-            match=data[0]['diffResults']['added'][2],
-        )
-
-        # Blacklists the two matching things, but not the middle one.
-        matches = rule.handle_log_entry({'data': data}, node)
-        assert matches == [expected1, expected2]
-
-
-class TestWhitelistRule:
-
-    def test_will_whitelist(self):
-        now = dt.datetime.utcnow()
-
-        data = [
             {
-              "diffResults": {
-                "added": [
-                  {
-                    "name": "good",
-                    "path": "/usr/local/bin/good",
-                    "pid": "12345"
-                  },
-                  {
-                    "name": "unknown",
-                    "path": "/usr/local/bin/unknown",
-                    "pid": "6789"
-                  },
-                  {
-                    "name": "othergood",
-                    "path": "/usr/local/bin/othergood",
-                    "pid": "10000"
-                  },
-                ],
-                "removed": "",
-              },
-              "name": "processes",
-              "hostIdentifier": "hostname.local",
-              "calendarTime": "%s %s" % (now.ctime(), "UTC"),
-              "unixTime": now.strftime('%s'),
-            },
-        ]
-        node = {'host_identifier': 'hostname.local'}
+              "id": "query_name",
+              "field": "query_name",
+              "type": "string",
+              "input": "text",
+              "operator": "equal",
+              "value": "asdf"
+            }
+          ]
+        }""")
 
-        rule = WhitelistRule(0, Rule.BOTH, config={
-            'field_name': 'name',
-            'whitelist': ['good', 'othergood'],
-        })
+        network = Network()
+        network.parse_query(query)
 
-        expected = RuleMatch(
-            rule_id=0,
-            action='added',
-            node=node,
-            match=data[0]['diffResults']['added'][1],
-        )
+        counts = defaultdict(int)
+        for condition in network.conditions.values():
+            counts[condition.__class__.__name__] += 1
 
-        # Whitelists the two matching things, but not the middle one.
-        matches = rule.handle_log_entry({'data': data}, node)
-        assert matches == [expected]
+        # Top-level AND, AND group, reused column condition
+        assert counts == {'AndCondition': 2, 'EqualCondition': 1}
 
-    def test_ignore_nulls(self):
-        now = dt.datetime.utcnow()
-
-        data = [
+    def test_will_not_reuse_different_operators(self):
+        # Different operators in top-level and group
+        query = json.loads("""
+        {
+          "condition": "AND",
+          "rules": [
             {
-              "diffResults": {
-                "added": [
-                  {
-                    "name": None,
-                    "path": "/usr/local/bin/good",
-                    "pid": "12345"
-                  },
-                ],
-                "removed": "",
-              },
-              "name": "processes",
-              "hostIdentifier": "hostname.local",
-              "calendarTime": "%s %s" % (now.ctime(), "UTC"),
-              "unixTime": now.strftime('%s'),
+              "condition": "AND",
+              "rules": [
+                {
+                  "id": "query_name",
+                  "field": "query_name",
+                  "type": "string",
+                  "input": "text",
+                  "operator": "not_equal",
+                  "value": "asdf"
+                }
+              ]
             },
-        ]
-        node = {'host_identifier': 'hostname.local'}
+            {
+              "id": "query_name",
+              "field": "query_name",
+              "type": "string",
+              "input": "text",
+              "operator": "equal",
+              "value": "asdf"
+            }
+          ]
+        }""")
 
-        rule1 = WhitelistRule(0, Rule.BOTH, config={
-            'field_name': 'name',
-            'whitelist': ['good'],
-            'ignore_null': False,
+        network = Network()
+        network.parse_query(query)
+
+        counts = defaultdict(int)
+        for condition in network.conditions.values():
+            counts[condition.__class__.__name__] += 1
+
+        assert counts == {'AndCondition': 2, 'EqualCondition': 1, 'NotEqualCondition': 1}
+
+    def test_will_not_reuse_different_groups(self):
+        # Different operators in each sub-group
+        query = json.loads("""
+        {
+          "condition": "AND",
+          "rules": [
+            {
+              "condition": "AND",
+              "rules": [
+                {
+                  "id": "query_name",
+                  "field": "query_name",
+                  "type": "string",
+                  "input": "text",
+                  "operator": "not_equal",
+                  "value": "asdf"
+                }
+              ]
+            },
+            {
+              "condition": "AND",
+              "rules": [
+                {
+                  "id": "query_name",
+                  "field": "query_name",
+                  "type": "string",
+                  "input": "text",
+                  "operator": "equal",
+                  "value": "asdf"
+                }
+              ]
+            }
+          ]
+        }""")
+
+        network = Network()
+        network.parse_query(query)
+
+        counts = defaultdict(int)
+        for condition in network.conditions.values():
+            counts[condition.__class__.__name__] += 1
+
+        # Top level, each sub-group (not reused), each condition
+        assert counts == {'AndCondition': 3, 'EqualCondition': 1, 'NotEqualCondition': 1}
+
+
+class TestBaseCondition:
+
+    def test_will_delegate(self):
+        class SubCondition(BaseCondition):
+            def __init__(self):
+                BaseCondition.__init__(self)
+                self.called_run = False
+
+            def local_run(self, input):
+                self.called_run = True
+
+        condition = SubCondition()
+        condition.run(DUMMY_INPUT)
+        assert condition.called_run
+        
+    def test_will_cache_result(self):
+        class SubCondition(BaseCondition):
+            def __init__(self):
+                BaseCondition.__init__(self)
+                self.runs = 0
+
+            def local_run(self, input):
+                self.runs += 1
+
+        condition = SubCondition()
+        condition.run(DUMMY_INPUT)
+        condition.run(DUMMY_INPUT)
+
+        assert condition.runs == 1
+
+
+class TestFunctional:
+
+    def setup_method(self, _method):
+        logging.basicConfig(level=logging.DEBUG, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+
+    def test_will_alert(self):
+        query = json.loads("""
+        {
+          "condition": "AND",
+          "rules": [
+            {
+              "id": "query_name",
+              "field": "query_name",
+              "type": "string",
+              "input": "text",
+              "operator": "begins_with",
+              "value": "packs/osx-attacks/"
+            },
+            {
+              "id": "action",
+              "field": "action",
+              "type": "string",
+              "input": "text",
+              "operator": "equal",
+              "value": "added"
+            }
+          ]
+        }""")
+
+        network = Network()
+        network.parse_query(query, alerters=['debug'])
+
+        # Should trigger the top-level alert, above
+        bad_input = RuleInput(result_log={
+            'name': 'packs/osx-attacks/Whitesmoke',
+            'timestamp': dt.datetime.now(),
+            'action': 'added',
+            'columns': {
+                'path': '/LaunchAgents/com.whitesmoke.uploader.plist',
+                'name': 'com.whitesmoke.uploader.plist',
+                # Remainder omitted
+            },
+        }, node={
+            'host_identifier': 'foo',
         })
-        rule2 = WhitelistRule(1, Rule.BOTH, config={
-            'field_name': 'name',
-            'whitelist': ['good'],
-            'ignore_null': True,
+
+        # Should *not* trigger the alert, above.
+        good_input = RuleInput(result_log={
+            'name': 'other-query',
+            'timestamp': dt.datetime.now(),
+            'action': 'added',
+            'columns': {
+                'a_column': 'column_value',
+            },
+        }, node={
+            'host_identifier': 'foo',
         })
 
-        expected = RuleMatch(
-            rule_id=0,
-            action='added',
-            node=node,
-            match=data[0]['diffResults']['added'][0],
-        )
+        alerts = network.process(good_input)
+        assert len(alerts) == 0
 
-        assert rule1.handle_log_entry({'data': data}, node) == [expected]
-        assert rule2.handle_log_entry({'data': data}, node) == []
+        alerts = network.process(bad_input)
+        assert list(alerts) == ['debug']
+
+        # Re-process the good input to assert that we don't continue to alert
+        # on good input after a bad one...
+        alerts = network.process(good_input)
+        assert len(alerts) == 0
+
+        # ... and that we *do* continue to alert on bad input.
+        alerts = network.process(bad_input)
+        assert list(alerts) == ['debug']
